+++ lib/helper/req_helper.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 import 'package:http/http.dart' as http;
 
-class RequestHelper {
+ class RequestHelper {
   static Future getRequest(String session, api, bool test) async {
     var url = test == true
         ? Uri.parse(
@@ -26,13 +26,13 @@ class RequestHelper {
     }
   }
 
-  static Future postRequest(String api, Map body, bool testMood) async {
+  static Future postRequest(String api, Map<String,dynamic> body, bool testMood) async {
     var url2 = testMood == true
         ? Uri.parse('https://uatcheckout.thawani.om/api/v1/checkout/session')
         : Uri.parse('https://checkout.thawani.om/api/v1/checkout/session');
     http.Response response = await http.post(url2,
         headers: {'thawani-api-key': api, 'Content-Type': 'application/json'},
-        body: json.encode(body));
+        body: jsonEncode(body));
 
     try {
       if (response.statusCode == 200) {