+++ lib/thawani_payment.dart
@@ -1,3 +1,4 @@
 library thawani_payment;
 
 export 'widgets/pay_button.dart';
+export 'functions.dart';